Add auth-db (#1)
Continuing the gradual migration of auth logic into saflib, this moves
the db portion. I moved the logic that was in
https://github.com/sderickson/saf-2025 over. Cursor agent did a bit of
refactoring to make the db not a singleton which I approved of, but
otherwise this is pretty much the same old implementation. I did take
the opportunity to add a test where, with the new instance
implementation, I spun up two clients on the same file and tested they
work.

## Documentation updates
My theory is documentation is the key to consistent output by agents. As
such, I'm filling out docs as I go based on what I needed to guide the
agent on for this PR.
* The agent tried to have the db library defer responsibility of data to
the consumer and change how the db behaved in testing. I'd rather
packages that manage state to store that state in `/data` when not
testing. Added `db-packages.md` with info like where to store `data` and
when.
* The agent at first would have exposed drizzle interfaces by this
library. Added `library-packages.md` with info like what functions to
expose, and how to expose types.
* Generally I had to nudge the agent a bunch on how to properly set up
the drizzle-sqlite3 setup. So I added a `setup.md` doc which hopefully
will work well when an agent creates a new such db package from scratch.
Will iterate of course...

### Unhandled agent misfires
What did the agent do wrong which I wasn't able to fix with
documentation? And how might I address them in the future?
* **Not exporting types**. This is a subtle issue, but if you export
functions and such for your library, but *don't* export their types or
return types or such, it's not a great dev experience. I might in the
future auto-generate library docs with `typedoc` which warns when that
happens.
* **Not fixing a mock**. This is probably the number one thing I see the
agent spin on, and I don't blame it. It needed to mock `argon2` in a
certain way to fix a test and it didn't know how to do it. Maybe should
add a doc on debugging mocking?

## Other work part of this PR:
* added unit test github action
* Updated gitignore
* Cleaned up tsconfigs, made the monorepo package the source of truth