Add auth permissions scopes (#3)
Add permissions to the auth-service, storing in auth-db, and propagated
as scopes to endpoints.

This is so that you can specify scopes on an endpoint and the client
needs to have that scope (via the user's permissions) in order to access
that resource.

Basically:
* You can configure `auth-service` with what permissions are possible
* Granted permissions are stored in an `auth-db` table.
* Whenever the user requests an endpoint, `auth-service` adds the user's
scopes as a header
* `caddy` propagates that header to the service that handles the request
* The openapi spec middleware now checks for scopes in addition to valid
request/response
* Routes can also access scopes themselves in the `req.auth` object

I've also added a test "ALLOW_ADMIN_SIGNUPS" where people can sign up
with an "admin" email. This is really for testing/demonstration, and
should not be used in actual production sites.

## Documentation

I added a great deal of documentation in this PR, as there was plenty of
opportunity:
* `auth-arch.md` - overview of how all this auth stuff works through
various components of the framework
* `queries.md` - for db queries. Updated based on back-and-forth when
setting up the user permissions query file
* `library-packages.md` - differentiated structure for service packages
vs library packages, as well as adding "config" folders to packages
* `ts-packages.md` - guidance on adding more packages. Hint: you need to
do this from the root dir
* `testing-middleware.md` - guidance on testing express middleware.
Trying to get more testing of the middleware in `@saflib/node-express`
* `setup.md` - for creating an openapi spec generation package.
* `query-testing` - frontend is where the agent messes up testing the
most. Trying to improve the success rate here...

What I *didn't* add documentation for is better component testing. It's
pretty tricky for the agent, and will often spin into trying to mock all
of vue query, which is a mess. I'm just going to have to deal with that
can of worms separately.