Add email verification (#6)
Mostly. Still no email system set up so no actual verification emails
can be sent.

This centers around creating the VerifyEmailPage component. This both
handles email verification links, and allows a user to send a new one
(if either the email verification failed or no token is in the url).

## Platform Updates
I made some heavy changes to the backend. For one, there's now a
function for creating an auth app, and that function is used in tests. I
also properly set up schema validation for the auth api and fixed a
bunch of issues there.

One larger change is tests no longer mock the db. I think it was overly
complicating things, and one of the benefits of sqlite is you can test
with it in memory. Documentation is updated.

I also finally found a way (well, prompted for a way) to enforce types
on endpoints. It's still not as automagic as I'd like (would like to
incorporate with `createHandler`) but one step at a time.

Another larger change is I broke down the large auth.ts router file into
individual route files. I was finding the agent was getting sluggish
making changes to it and the test file, so keeping files generally below
500 lines I think is a good target.

There are also more helpers I want to make for testing api handlers but
the auth service is sort of a special case, so I'll wait and see what I
want to port over later. JIT-framework building.

I think I might want to just delete tanstack query handler tests,
though. Or, at least rework them to test how they interact with each
other. The component testing covers the request handlers pretty well.

Still tuning the spec template. I found the original spec to a bit too
prescriptive, that it confused the executing agent.