Add security to auth-service (#8)
Buncha stuff to the auth service:
* session secret auto creation, storage, and rotation
* only trust caddy proxy
* tighten tests, add thorough coverage
* rate limiting
* tighten error response schema

Improvements which will include all "node-express" services:
* csrf protection
* helmet
* cors no longer includes localhost domains by default

I asked a thinking model to audit the service and provide suggestions. I
implemented the more important ones like CSRF and rate limiting. Other
things I might do in the future to upgrade things:
* Session invalidation when user changes password (should set up a
change-password feature first...)
* Rate limiting generally for other APIs, but I want to leave that to
the individual service to decide
* RBAC
* More env variable configuration

The one main thing the model mentioned is missing is logging, anomoly
detection, and alerting, but I need a logging and alerting services for
that. Once that's done, though, I consider this good for standard use in
production. As-is, it's still solid enough for small use.

Other improvements
* add "createSafClient" for vue-query. Auto-adds CSRF token