Add emails to auth-service (#11)
Not much of an auth service if it can't send password reset and
verification emails.

This also:
* adds a generic "@saflib/email" package which can be configured to send
to any SMTP endpoint via nodemailer
* this includes a custom mock so that consumers of the library can run
`vi.mock("@saflib/email")` and it'll Just Work. I'm trying this out but
if this works well I might copy it for other libraries which make make
external calls when they're used in unit tests

Also:
* updated docker.ts to handle bun. I've got a service I'm working on (to
be added to SAF later) which uses bun, and it needs all package.json
files added to the image for it to work :/
* some updates to `ts-packages.md`. I tried to iterate on it a bunch to
get agents to more consistently create new ts packages but I was only
partly successful. My main takeaway is Gemini is more consistent than
any of the other standard-cost agents available in Cursor, but still not
super reliable.