Add basic admin SPA (#12)
Now that there are scopes, and verified emails, I feel comfortable with
having an admin interface. This change gets things started by:
* adding a user list endpoint
* creating a new "admin spa" which allows verified admins to view the
users

As part of this:
* I had to refactor auth-service. It's handling now the "/users"
endpoints as well as the "/auth" endpoints, and the latter require a
bunch of custom middleware (like CSRF, sessions, and rate limiting) that
the "/users" endpoints don't need.
* As part of that I moved the middleware into a `middleware` folder
* I also started moving the scope system to be more RBAC. That is, there
isn't an "admin" scope, since an admin is a role, instead I'm adding
scopes like "users:read" and auto-granting those scopes to admins. A
formal setup there isn't yet, but I'd like the auth service to take a
config. One day.
* Changed up the email mocking system to not break when a test clears
all mocks. Got rid of some less-important tests that were depending on
the old implementation.
* Made some of the middleware match the common error response structure.
Need to figure out how to share that better, though...
* Improved the scope middleware to use the conveniently-added `openapi`
property on the `req` object. Rather than try to find it itself.
* Fixed the `stubGlobals` to get rid of retries by the tanstack client,
so errors show up immediately and are easier to test.
* Fixed up some docs

Also, I considered putting the new user admin pages into a separate
package, but I opted not to. The reason is this: I think the ownership
should be centralized. I'm thinking of packages as units of ownership,
and it makes sense for the team that owns identity frontend to also own
internal-use tools. Similarly, I think common components which are
centered around the /users apis should live in this package, even if
they might live on other pages.