Various updates 4/16/25 (#14)
I'm working on a cron job service in a separate repo, and these are
various changes from doing that work.

* Use `locals` for accessing database in auth-service. I realized the
way I'm doing it is polluting the "type" namespace. So I'm going this
route instead. I still need to come up with a better solution for better
typing express route handlers...
* Start adopting a common "error" object for apis. I have mostly been
having apis return an object with an "error" message but it seems more
common is an object with a "message" property. I wish there were a
better way to access the error schema through npm packages but for now
I'm using relative paths. There's more adoption and updating to do...
but that's for another PR.
* Fix tests from a tweak earlier where generated docker files now have
consistently sorted packages.
* Some doc updates.
* Add a "node-logger" library. Eventually I'll have this logger send to
Loki or some-such, but this already helps in that it streamlines testing
logs in integration tests. See "logger.test.ts"
* I'm setting "isolate" to "false" for vitests. I didn't see much change
in performance... nor did anything break. But I'd rather turn that on
now and fix issues that come up with it as I go.