Refactor spas, add some ts:checks (#16)
Working on a better way to organize vue into packages. As part of that,
I've added `createVueApp` to `vue-spa` so each app doesn't have to
re-implement the basic setup.

I'm also trying out adding some ts checks. They're breaking, but for
small stuff. Will get those fixed and auto running in CI in a later
commit.

And also made `auth-vue` export a "createAuthRouter" which creates and
returns a singleton. Having a function create one makes it so that
importing `auth-vue` has no side effects (it was competing for control
of the history of another router which was running) and the singleton is
so that tests don't get messed up.