Bunch of backend software design updates (#17)
I'm working on building out a backend, and the structure I have now is
getting a bit much. I'm finding that to manage things like routes and
database queries, it's easier with AI to do so when the files are
smaller. So I deconstructed singular files with multiple routes and
queries and updated docs here to match.

I've also adjusted how backend packages handle errors. As a general
rule, they should return an object `{ result }` or `{ error }`, and use
the new provided `ReturnsError` generic type. That way consumers of
these functions are nudged with TypeScript to check for errors and
handle them. Ideally with a `switch(true)` statement which defaults to a
TS `never`.

```
  const { result: callSeries, error } =
    await ctx.db.callSeries.get(callSeriesId);

  if (error) {
    switch (true) {
      case error instanceof CallSeriesNotFoundError:
        return res.status(404);
      default:
        throw error satisfies never;
``` 

Then if the initial implementation doesn't handle all error types, TS
will flag it.

I'm also starting to work with data which relates to each other. For
that I'm putting in documentation encouraging flattened data structures
when serialized.

Also started adopting `AsyncLocalStorage`. Just for the db for now, but
a winston logger will also be a fine addition.

These changes generally up the quality of backend code written in SAF,
especially when there's relational data and non-trivial endpoints.