2025 05 06 updates (#23)
Bunch of node-related changes:
* @saflib/node-express and @saflib/node-logger have been refactored to
@saflib/node and @saflib/express.
* Added a shared "context" to @saflib/express - a place for a logger, an
auth object, and a request id
* Reworked the express middleware, in particular got rid of (most)
putting-of-things on the request object, in favor of the context object.
"auth" I left there, marked deprecated, because I think that's a bit
more widely used.
* Updated `auth-service` to use the new context logger
* Added some documentation on testing database queries, and updated the
workflow
* Some work I'm going to throw out on grpc-servicer...
* Added a workflow and updated docs for creating ts packages.
* Added a typescript override saying ImportMeta really does have
dirname, filename
* Removed "vitest.config.js" from coverage reports

I'm basically working on 1st-class support for gRPC between Node
services, and the context stuff is a big part of that. I'm making it so
that any logic which is triggered by either a Node/Express request or a
gRPC request will have basic stuff in Node AsyncLocalStorage contexts.