Add xstate-based workflows (#27)
Following what I wrote
[here](https://scotterickson.info/blog/2025-05-10-Workflow-First-Iteration),
in this PR I'm trying out xstate to provide a bit more structure and
features to dev workflows. I picked xstate in part because claude keeps
recommending it, and also because going through the docs it seems pretty
comprehensive.

I've already started using it for some backend process stuff, so this PR
is taking what I learned there and extending it here. I might create an
"xstate" saflib package at some point to colocate logic which is useful
across different use cases (such as "allChildrenSettled").

In lieu of porting all existing workflows to the new model, I added
abstractions so xstate-based workflows and the prototype workflows have
the same interface which the CLI uses.

I did port one though: the "add tests" sample workflow. In addition I
tried using some of the xstate features such as guards and raising
events. Based on this I kind of see there being largely two types of
states:

* **Potentially Skipped Step**. In this kind, there's an `invoke` which
does a check and, if successful, moves to the next state in `onDone`. If
not, the `onError` prompts the agent to do something about it.
* **Always Executed Step**. This is a core part of the workflow that
will always run. This prompts the agent on entry and only moves forward
if explicitly signaled to do so.

I *may* create factories for these sorts of steps just to make building
workflow state machines more ergonomic but for now I want to make a
handful more manually before creating a bunch of generic tools.

The types were the hardest part of this work. I'm not 100% sure I set
these up (esp in `xstate.ts`) the way they should be, but in *using*
this shared logic I'm getting the experience I want at least...