More workflow improvements (#31)
Continuing to update workflows as I work my through creating a website.

New and updated workflows:
* `update-impl`: a generic workflow for updating an *existing*
implementation. This I think can be more generic than other workflows
which are context-specific. It worked so far on updating express handler
implementations, will see how it handles pointing it at other things.
* `spec-project`: updated to xstate, and extended to include creating a
checklist. Uses the `saf-doc` tool for that purpose.
* `add-spa-page`: Added *gradual* implementation steps. My approach to
pages is not the natural way agents build frontend, so I'm trying
spelling it out, step by step. Also improved the template files to push
agents in the right direction.

One other thing I built especially for the "spec-project" workflow is a
"saf-doc" command which lists all the packages in the workspace and
their descriptions. I see the agent doing a lot of navigating around the
codebase which I *think* is unnecessary, and can be mitigated by
providing high-level or context-specific information. Anyway, you can
run `npm exec saf-doc` anywhere and get a nice list of packages, and
then `npm exec saf-doc <package-name>` will be used for providing more
information for that particular package. The spec-project workflow calls
this and provides the output to the agent.

Other improvements:
* modernized the users tanstack query
* made the workspace.ts logic find the nearest monorepo root
automatically
* some package descriptions so `saf-doc` prints full info
* fix for `update-spec` workflow
* changed the loader function signature from an array to an object,
since typescript likes that better when there are multiple things
loaded. Otherwise the auto-type is an array of mixed types, rather than
a tuple which was the intention.
* standardized workflow param names