Auth service grpc, workflow updates (#32)
Bunch of product-related changes.
* Added a grpc service to the auth-service which handles one RPC:
getUserProfile. This is also the first example of a service running two
servers at once: grpc and express, as well as the first example of a
service running a grpc service with the same context as an express
server. It seems like this pattern *might* be common.
* Updated workflows add-queries, update-schema, add-route, spec-project
* Added workflows add-email-template, add-grpc-server, add-grpc-handler,
add-proto-package, update-grpc-spec, add-node-xstate
* Refactored a wrapping function into grpc-node to simplify working node
async local storage into handlers, as well as some testing helpers to
run grpc in tests (in particular when using vitest fake timers, which
cause the grpc server to hang)
* Tweaked the output in workflows, trimming lines, indenting added
lines, and adding another utility function
* Added a helper method to the workflows package, for creating a
prompting state, used it in add-route.

Following [my post on accountability and
gaslighting](https://scotterickson.info/blog/2025-05-24-Accountability-and-Gaslighting),
I think I need to tighten and build out my workflow tool library. Much
of this repository is really just "glue", helper methods to bring
together and use various technologies in the stack, but this workflow
tool might make more sense as a standalone package, as it brings the
accountability to the agents that I think is so necessary.

Main thing to do now is to create helper methods for the common
behaviors. Running checks, copying template files, and prompting to read
docs. Beyond that, this might make sense as a full-fledged server which
runs locally... Big plans for this thing.