Add cron service, add auth service interface (#38)
I've been building and using a cron service for a product, but I'm
finally moving it into SAF where it belongs. Mainly I just hadn't
decided how to export services in a reusable manner. I did double duty
here: now both the auth service and the cron service, reusable services,
are in SAF, and products can import their main functions and pass in
whatever options. For auth, it will be interfaces for account creation
callbacks and email templates, for cron it's the job definitions.