Add utils for managing test artifacts (#62)
Working on making it easy to view artifacts generated from automated
testing. This involves:
* adding a check to auth-service so Caddy files can make it so only
admins can access these artifacts, but also not hide them behind CSRF
checks
* adding a "saf-test-assets" bin command to generate coverage reports
and copy all of them and all e2e tests with screenshots and a manifest
file into a target folder.
* adding a "test-coverage" bin command to make it easy to run the tests
either normally or piping stdout to a file to go in with everything else
in the coverage folder