Remove 401s from /auth/verify (#67)
It doesn't make sense for the auth service to know and enforce all the
routes. So it no longer returns 401 whenever it runs and there isn't a
user, it just doesn't return any user information.

The responsibility is now on the auth middleware. I moved it in the
order of middleware so that it has access to the openapi route schema,
and can check if there's a "no-auth" tag. This maintains the
"on-by-default" behavior where if you just use the recommended
middleware with default settings, routes return 401 unless the user is
logged in.

Also tweak some playwright helpers. "v-text" never worked.

Also removed playwright helper unit tests... there aren't any there, but
they're failing in CI. But they're not doing anything so removing them
for now.