Mock email lib (#68)
Creating a general system for mocking emails in testing and development,
and viewing them. The main value is getting to screenshot and test the
links in emails in playwright tests. Then I can test, for example, the
entire email validation flow.

As part of this, I changed how I mock the email library. Rather than
using the vitest mocking system, the ability of the library to turn on
mocking is a full-fledged feature. I've been trying this with a couple
third-party services and I'm keen on it.

So, the further I go, the more I wish each client and each server had
its own subdomain. This also comes to play with these email endpoints. I
will change that later...

Oh, I also added a small middleware that just blocks html input.
Rendering even data in a mocking, sandbox context made me a bit nervous,
and figured might as well add some protection on the API end against
HTML arriving through most normal interfaces.