Update versions, typecheck everything (#70)
Started by trying to finish #42. Scope increased to also updating
everything to latest versions.

They typechecking is using `vue-tsc`. That works for non-vue code, as
well as vue code. It's not the most elegant solutions; ideally I think
everything would just be checked with `tsc` but it's an improvement from
before.

I'm also a little unsure what to do with the tsconfig.json files in all
the packages, and the individual package `typecheck` commands. It's
inefficient to run each of them separately, but there might be cases
where it's good to be clear what the typechecking config should be for
any given package, and to focus on it and its non-built dependencies. I
kind of wish there was a way to "scope" typechecking to a specific
folder, but I think TypeScript and the ecosystem is sort of built on the
assumption that you build every package separately, which I'm not doing
here. Anyway, I'll leave the individual tsconfig files for now, and
consider what to do with them later.

Renamed "test assets" to "health assets". I was thinking about reporting
whether a package is typechecked or not and have that be a metric, but
perhaps it should just be expected to always be true. Anyway, I kept the
rename.

I should figure out a story for types and tests that are in workflow
template files. For expediency I just dropped in some `// @ts-nocheck`s
in there with a TODO to remove them so agents will have to to finish a
workflow, but it would probably be good if types still were checked so
template files could be updated easier and sooner. Workflow tool needs
work generally.

I'm not sure how express ended up on version 4. Fixed that.