Add grpc metrics (#78)
Filling out metrics to send to prometheus, this time for grpc. Pulled
out a grpc "runner" which handles errors, and now also tracks duration.
And I added a function to have prom-client collect system (default)
metrics.

Also:
* Do some refactoring of how grpc services get added. I swear I remember
trying to subclass the abstract service class before and typing not
working or whatever, but now it works 🤷
* Move prom-client to the node package, since it's generally for node