Move clients to subdomains (#80)
I've been bothered by SPAs as sub-paths of a given domain, where it
makes more sense and more clearly organizes things to have each spa be
its own subdomain. So I'm finally doing something about it. Mainly on
this side of things, I just updated the auth code, and the tanstack
client code.

As part of this, I also centralized links for spas. Rather than
hardcoding each link in a bunch of different places, I'm creating
dedicated packages with "Link" objects that can be used to create intra-
or inter-spa links, either on the frontend, backend, or in tests. This
vastly reduces the danger of broken links. I added a "@saflib/links"
package for shared logic there.

I also made a new env variable: `CLIENT_SUBDOMAINS`, in particular for
CORS.

I plan on doing something similar for services.

This is a good thing to do now because how I organize clients and
services and their paths will relate to how I want to organize the code
(#47), how specs and rpc files get organized (#46) and how docker files
get generated (#44). Working on a general mental model, but I like
having the top-level application breakdown be 1:1 to subdomains.