Add @saflib/env (#83)
It's been bugging me for a while how little visibility one has into what
environment variables a service or package depends on or expects or has
available to it. So I'm adding a simple system for defining and typing
env variables.

The core of it is the new `@saflib/env` package, which provides:
* A CLI tool for seeing env variables a package has, based on
`env.schema.json` files in that package and its dependencies (`npm exec
saf-env print`)
* That same tool also can generate combined schema files and typescript
types (`npm exec saf-env generate`).
* A function that services can use to get a validated, typed env object.

It also defines some key env variables which are presumed to be
available everywhere:
* `DOMAIN` and `PROTOCOL`, as I expect everything to be served through
one of these together.
* `CLIENT_DOMAINS` - what subdomains are user-facing websites
* Added a `DEPLOYMENT_NAME`. Right now just to make it clear what
deployment a given database is for, but also should probably include it
in things like instrumentation later. This also starts moving things
away from using `NODE_ENV`.
* `TZ` - forcing it to always be `UTC`
* `NODE_ENV` - restricting its values
* `MOCK_INTEGRATIONS` - used by various libraries which make calls out
to 3rd party services. In development and in testing, this should
usually be true.
* `ALLOW_DB_CREATION` - a safety override I set up to make sure changes
in database naming (accidental or otherwise) doesn't end up creating a
whole new database.

I use JSON schema at the core to define and validate the data and
generate the types. I started to use fancy features like `allOf`, but
this produced overly complicated types and combined schemas. So I'm sort
of restricting the types of schemas that can be defined. This neuters
much of the power of json schema so... there's probably a better
solution out there but for now this is very much in the right direction.

I made some improvements to what's included in the environment, mostly
prepending service names to env variables to denote more clearly how
they're namespaced.

This tool could certainly use improvements, or more ideally, it should
get swapped out with something else. But this is the minimum feature set
I'd need: definition, validation, and type-checking.