Generalize nodemailer config (#87)
Rather than have a whole bunch of env variables for the nodemailer
transport, I'm putting a json object into a single env variable and
putting that in.

Also, I'm trying out more being able to override env variables in
development. To test that update email configs worked, if the env
variable is set that overrides "MOCK_INTEGRATIONS" so that you can test
a live integration in a targeted manner. I'd like a more secure approach
in the future, but something like this I think would be useful for both
manual and automated local testing of integrations.