Simplify "subsystem" (#90)
I was thinking about changing this, at least to have terms that are more
subsystem-specific (instead of operationName, it would be routeName,
methodName, etc).

But let me try this first: restricting "subsystemName" to a specific
list of options. I do think I need some concept of "subsystem" or a name
for some high-level portion of a service. I might change operationName
later but for now I think this simplifies things nicely.