Introduce env workflow, improve workflows (#92)
Having just added typed envs, there's a process now for adding new
variables. To get that working, I:

* Added a `generate-all` command to `saf-env` CLI, which goes and finds
all existing `env.ts` and `env.schema.combined.json` files and updates
them. Many of the changes in this PR is just getting the generated files
updated.
* Updated the "add-workflow" workflow to the new setup, to be something
like the "add-spa-page" workflow. I'm still playing around with how to
make the workflows more readable but the factory pattern seems to be the
right direction. The template is also updated to the factory pattern.
* Added the `add-env-var` workflow also with the factory pattern.
* Updated the node logger to not have a format by default, and updated
the workflow cli to adopt it. This will make logging simpler within
these machines.
* Tweaked the spa page workflow to adopt the latest changes,
particularly i18n and link strings.
* Added a factory for running npm commands, with an explicit list of
allowed commands. It doesn't provide much safety right now because the
agents can run whatever commands they want outside of the CLI tool but
it's good to have a tool like this be built to be limited.