Mock identity-rpcs (#93)
Trying out building more mocks for consistent unit testing. To that end,
I've made it so the identity service, if you use the default client,
will return mock data rather than trying to connect to a grpc server.

Basically I'm leaning toward the idea that it's the job of an SDK or
client interface to provide a decent, mocked experience so those that
depend on that service can unit test things fairly easily without having
to create and maintain their own mocks. In the case of SAF, that means
tanstack query libraries, grpc libraries, and integration libraries.