Add verify wall page (#94)
Add a page where users can be sent after registering which requires they
verify their email. It's up to the application to set up the register
and login pages to redirect to it, though, and to add any further
redirects which will send unverified users to this page.

In addition:
* hard-coded a copy of the error object in the cron spec schema. I just
need a better system for sharing static assets which aren't javascript
* adjusted the auth router options so they can take a default redirect,
and so that tests don't complain about a missing "redirectUrl prop"
* modernized the auth-vue page unit tests, to use a common test app.
* made the verify email page emit a "verify" event so a wrapper page can
choose what to do (in my case, auto-redirect), and also handle better
when the user is already verified (it shows success, rather than trying
to submit the old token again)
* added a missing "declare module" bit in auth-vue/requests/client.ts
which makes sure the tanstack error is typed correctly
* namespaced the default auth client strings