Document package workflows (#106)
This doesn't document `@saflib/workflows`; it updates `saf-docs
generate` to produce markdown files for the workflows of the current
package.

In addition:
* Added a high-level workflows doc.
* Refactored a couple cli tools to be multiple files.
* Generated docs for `add-env-var`
* Improved workflows logic some: more helpers, added a source url and
example values, and an input interface
* added env to the workflow-cli
* refactored the factories for copying and updating templates in
workflows, breaking it down
* added "dry runs" and checklists to the factories that add-env-var uses
* added checklist, list, and source commands to workflow cli
* renamed "useTemplateStateFactory" to "copyTemplateStateFactory"

This is plenty. The two major follow ups that still need to happen are:
* Updating existing workflows to XState
* Generating docs for each of those

I'll work my way through generating workflow documentation for each
package next.