Document @saflib/node (#109)
More complex than I realized. Wrote up an instrumentation doc that
breaks down the provided logic into these use cases: * using
* integrating
* providing
* metrics

I figured out that when I made `saflib` no longer a flat list of
packages, that the root vitest command (and by extension, CI) is no
longer testing nested packages like in `identity` or `cron`. I fixed
that, and fixed a few tests that broke along the way.

I decided to change logging a bit to make it better for testing, and to
reduce the amount of things exported by `@saflib/node`. Now
`getSafReporters()` and `getSafContext()` return full stores when
`NODE_ENV` is `test`. This way it Just Works in most cases, and in
specific cases where you want to check the logs in a test (like in
`@saflib/cron`), the test can get the same objects and spy on them. I
also made it so that errors which would normally be sent to console in
tests can be silenced by having the test use `addErrorCollector`, on the
presumption that if that happens in a test, the test is expecting it to
happen.

Originally when I built the `drizzle/sqlite3` wrapper, I hadn't settled
on the `ReturnsError` best practice, so it would expect typed errors to
be thrown. Since I adopted `ReturnsError` repo-wide, the db
`queryWrapper` no longer makes sense to have catch and discern between
handled and unhandled errors. Anything that's thrown is unexpected and
unhandled; handled errors are returned. So I got rid of that bit of code
in there.

I moved `metricHistogramDefaultBuckets` out of `@saflib/monorepo` and
into `@saflib/node`. Makes more sense there.

I ended up needing to provide `saf-docs generate` with context (not
exactly sure what changed there to need it but c'est la vie), so I added
the `@saflib/commander` package, moved the context/reporter setup in
there, and added it to existing CLI commands.