Overhaul workflows (#114)
Having moved all workflows to an XState-based structure, I decided I
should keep the power of XState but hide the complexity. This PR gets
rid of the `ConcreteWorkflow` as the workflow "unit" to export, and
instead uses an object that follows a `WorkflowDefinition` interface.

The main value to this is it's much easier to read and maintain a
workflow, and to see what are the main elements. It also makes the
package interface muuuuuch better.

This PR then has:
* Updated all workflows to use this new definition format.
* Added tests which dry run each of them.
* Cleaned up and reorganized the workflows package.
* Updated the workflow package's README
* Namespaced and adjusted some workflow names

Some other changes:
* Prettier'd some files
* Added another global that ended up being used in vue testing
* Added a "generate-all" command to "saf-docs" CLI, and ran it to update
all docs.