Publish saflib-workflows (#116)
Right.

So having overhauled this library in #114, this gets everything ship
shape for npm.

* Add LICENSE
* Tested the add-workflow to make sure it works by adding drizzle/init
workflow 🔥
* Add CHANGELOG
* Remove all internal dependencies, and some external dependencies, from
the workflow library.
* Overhauled logging as a result because winston was messing things up?
Also seemed like overkill.
* Move some workflow machine guards to setup, because apparently types
mess up packing otherwise. XState types are...
* Add a custom "lineReplace" input for the copy machine, specifically
for adding a workflow to deal with the public package having a different
name than my internal one (more on that later).
* Replace "getGitHubUrl" which was creating urls to my repo, with
"getSourceUrl" so other workflow CLIs can link to their own repos.
* Export some types so build work.
* Make sure the README works on npm's site.
* Add the whole publishing mechanisms.

In terms of packaging, I decided I wanted to keep the tsconfig and
package for distribution separate from the one I use internally. This is
so that:

* Intellisense takes me to the TypeScript file when I follow a ref to
the workflow library, rather than a compiled JS or types file.
* Trying to get my preferred development rules, and the build rules, to
be the same was no bueno.
* I'm using private namespaces for my repo, but I'm not convinced that's
worth $7/mo to publish so I'm publishing with a public name instead.

So the "dist" folder in the workflows is the distributed package. The
process for publishing is to run `npm run pack` in the workflows , then
`npm publish` in the dist folder. I did this on the latest commit on
this branch, then I'll merge it in and tag/release that commit as the
same thing (though not technically? I'll sort out a more well defined
release process as I go).

🚢