Workflows v0.1.2 - Checklist fixes (#124)
I realized if you run `npm exec saf-workflow checklist` the prompts
print out. It was also breaking if you go ahead and use a workflow
machine within another workflow. I fixed the prompt step machine to
continue on in a dry run.

Also, in playing around with having workflows in workflows, I realized I
decided to have the machine return the wrong thing for checklists. It
should return a single item, not a list. Fixed that, but that doesn't
affect anything on the public interface.

As part of building toward workflows within workflows, I added another
property to `WorkflowDefinition`: a `checklistDescription` which will be
based on the context, and decide what gets printed out when the workflow
is part of another workflow. If none is provided, it will fallback to
`description`.