Rework commander and drizzle workflows, add composite workflows (#125)
I've pretty much decided the best way to organize template files is to
have all the workflows in a package draw from the same set. The
"templates" directory they all draw from will be organized in the way a
dependent package ought to look. This has a few benefits:
* The "templates" folder gives a holistic view of what a dependent
package ought to look like
* The files themselves can be typechecked, since singular file templates
(like a package's index.ts or tsconfig files) exist relative to the
templates in the correct place.
* When copied over, the imports and types still tend to work (as long as
where they're being copied into follows the same directory/file
structure.... which it should).

So I started reorganizing the commander and drizzle workflow templates
accordingly. I found myself really wanting to test them though; I knew
there would be runtime and process issues which would not or could not
easily be caught with static tools. And I've also been wanting to
combine workflows into one-off "plans" which aren't necessarily listed
in the workflow tool.

So I extended workflows to be composable.

The examples of this are the "test-workflows" files in this PR. There's
one testing all the commander workflows, and another testing all the
drizzle workflows. Sort of an E2E test for a package's interrelated
workflows. These discovered some logical and process issues so I updated
things accordingly. I also added some checks for the logical issues,
like checking that a "fileId" given to the "UpdateStepMachine" is a
valid one.

I had to make some fixes for composite workflows to work:
* **Added `cwd` to the context**. As well as a `CwdStepMachine`. And
updated all the workflows to use that instead. Now you can have as a
step in your workflow to change the current working directory, and it
will affect the rest of the steps and all sub-steps. It's always
relative to the process's current working directory, which I may want to
change later...
* **Added `kickoff-unlisted` command** to workflows CLI. This way I can
run a workflows file without incorporating it into a CLI, and is
especially useful for workflows which are tests or one-off plans.
* **Added pollingWaitFor**. A function to replace xstate's `waitFor`
because that won't stop Node from exiting the program before the machine
has finished. Hopefully I can figure out a proper way to set up these
machines so a polling wait isn't necessary in the future. Mostly used
internally but figured I'd export it for any other runners.

Also:
* Updated docs for these workflows.
* Fixed the github links in generated docs.
* Moved the workflows CLI logic into a bin folder to better match the
commander layout
* Some logging improvements